move factory methods to new file
